+++ accounting/reports/report.go
@@ -5,9 +5,9 @@ type TrialBalanceRow struct {
 	AccountCode string
 	AccountName string
 	Category    string
-	TotalDebit  float64
-	TotalCredit float64
-	Balance     float64
+	TotalDebit  int64
+	TotalCredit int64
+	Balance     int64
 }
 
 // LedgerRow es una fila del libro mayor de una cuenta.
@@ -15,9 +15,9 @@ type LedgerRow struct {
 	JournalID   string
 	Date        string
 	Description string
-	Debit       float64
-	Credit      float64
-	RunningBal  float64
+	Debit       int64
+	Credit      int64
+	RunningBal  int64
 }
 
 // IncomeStatementRow es una fila del estado de resultados.
@@ -25,21 +25,21 @@ type IncomeStatementRow struct {
 	AccountCode string
 	AccountName string
 	Category    string
-	Amount      float64
+	Amount      int64
 }
 
 // IncomeStatement agrupa ingresos, gastos y la utilidad neta.
 type IncomeStatement struct {
 	Revenue     []IncomeStatementRow
 	Expenses    []IncomeStatementRow
-	NetIncome   float64
+	NetIncome   int64
 }
 
 // BalanceSheetRow es una fila del balance general.
 type BalanceSheetRow struct {
 	AccountCode string
 	AccountName string
-	Balance     float64
+	Balance     int64
 }
 
 // BalanceSheet agrupa activos, pasivos y patrimonio.
@@ -47,20 +47,20 @@ type BalanceSheet struct {
 	Assets                    []BalanceSheetRow
 	Liabilities               []BalanceSheetRow
 	Equity                    []BalanceSheetRow
-	TotalAssets               float64
-	TotalLiabilitiesAndEquity float64
+	TotalAssets               int64
+	TotalLiabilitiesAndEquity int64
 }
 
 // CostCenterRow es una fila del reporte por centro de costo.
 type CostCenterRow struct {
 	CostCenter  string
-	TotalDebit  float64
-	TotalCredit float64
-	Balance     float64 // debit - credit
+	TotalDebit  int64
+	TotalCredit int64
+	Balance     int64 // debit - credit
 }
 
 // CostCenterReport agrupa las filas y el balance neto.
 type CostCenterReport struct {
 	Rows    []*CostCenterRow
-	NetDebt float64 // suma de balance positivo (gastos netos por centro)
+	NetDebt int64 // suma de balance positivo (gastos netos por centro)
 }
